+++ src/Bavfalcon9/Mavoric/Cheat/Movement/SpeedA.php
@@ -1,0 +1,35 @@
+<?php
+/***
+ *      __  __                       _      
+ *     |  \/  |                     (_)     
+ *     | \  / | __ ___   _____  _ __ _  ___ 
+ *     | |\/| |/ _` \ \ / / _ \| "__| |/ __|
+ *     | |  | | (_| |\ V / (_) | |  | | (__ 
+ *     |_|  |_|\__,_| \_/ \___/|_|  |_|\___|
+ *                                          
+ *  This program is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU Lesser General Public License as published by
+ *  the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ * 
+ *  @author Bavfalcon9
+ *  @link https://github.com/Bavfalcon9/Mavoric                                  
+ */
+namespace Bavfalcon9\Mavoric\Cheat\Movement;
+
+use pocketmine\Player;
+use pocketmine\event\Listener;
+use pocketmine\event\player\PlayerMoveEvent;
+use Bavfalcon9\Mavoric\Mavoric;
+use Bavfalcon9\Mavoric\Cheat\Cheat;
+use Bavfalcon9\Mavoric\Cheat\CheatManager;
+
+class SpeedA extends Cheat {
+    public function __construct(Mavoric $mavoric, int $id = -1) {
+        parent::__construct($mavoric, "SpeedA", "Movement", $id, false);
+    }
+
+    public function onPlayerMove(PlayerMoveEvent $ev): void {
+        $player = $ev->getPlayer();
+    }
+}
